requirements/robot.txt
aiogram~=3.18.0
python-dotenv~=1.0.1
ujson~=5.10.0
cachetools~=5.5.2
tortoise-orm[asyncpg]~=0.24.0
# tortoise-orm[aSQLite]
tomli_w~=1.2.0
tomlkit~=0.13.2
aerich~=0.8.1
